squelch another 'set but not used' variable warning.
I worry a bit, though, that there might be a real problem here -- currently the parent of the photon from nuclear de-excitation is the initial-state nucleon/nucleus.  This seems a bit strange: it's not like the struck nucleon in a CCQE interaction (for instance) is really the thing decaying.  But I'm not sure what else one would call it, given that the post-interaction pre-excitation nucleus isn't always part of the particle record...

git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/genie/generator/trunk@6366 cc9776de-3512-45ca-aafc-e2d9ed43c22c