Nuclear & Interfaces directories had no Makefile (and weren't being visted by top level Makefile) so their headers weren't being "installed".
Nuclear/LinkDef.h needed to be told explicitly to pull in the headers
that it was making dictionary for but aren't in that directory
(Types/NuclearModel.h and Interfaces/NuclearModelI.h)

make/Make.include deconstructs the ROOT version so we can decide rootcint
vs. rootcling (w/ pcm and rootmap support).

make/Make.std-package-targets has a much more elaborate handling for ROOT 6
dictionary building.  Lots of -I include paths given so that dictionaries
can find their headers where GENIE is used in build situ, installed or
bundled up and relocated as a UPS product.



git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/genie/generator/trunk@6500 cc9776de-3512-45ca-aafc-e2d9ed43c22c