Feature: added Helium, to be generated using the Carbon hadron tensor and the regular scaling function. Also let the code go down to Be9, a default GENIE isotope.
Activated lines in pdg table for two pseudonuclei 
made of pp and nn, so that generation on Helium can proceed
but didn't activate (kept commented out) the decay modes
because it borked the free-n line that followed.

Bugfix: in the scaling function that 
affected calculation for the neutrino case.
Bug only affects nuclei scaled away from a hadtensor.
No effect on C, O, Ca40, Pb.  
(Bug not in pre-2.12 devel versions, so MINERvA Pb,Fe ok.)

Fixed another bug that set a too-low XSecMax
for O and maybe N14.  XSecMax more generous, safe.

The use of NucBindE was disabled (?) for 2p2h in 2.12
so switched and added the resulting removal energy 
directly to the remnant nucleus.

Sara B's q0 qz table filling switch was already in.


git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/genie/generator/branches/R-2_12_6@7035 cc9776de-3512-45ca-aafc-e2d9ed43c22c